US8670: Authentication based home page delivery
This commit adds the functionality to display two different CMS pages to
the "/" path based on a user's authentication status. This is in support
of feature F270 (Homepage Redesign) in which it is the intent to deliver
a custom content, dashboard like experience to users logged into the home
page.