US10385 Dynamically display next stream date
1. Contextualize any major architectural or UI changes in this PR.
Previously the Date and Time displayed in the jumbotron overlay for the
next event that was greater than 15 hours away was hard coded HTML. The
changes in this commit dynamically inject the next streamspot event data
into the DOM. In order for this to work correctly, there is an
associated code change in the CMS repo for removing the static HTML.

2. Do these changes belong here, or somewhere else?
Here

3. Does the code follow our coding standards and practices?
Yes

4. Has the code been linted?
Yes

5. Have you written tests for the code you wrote / changed?
Yes

6. Have you run your tests locally?
Yes

7. Is there anything that feels like a dirty hack? If so, have you worked with
someone to find another, cleaner way to accomplish the same thing?
In order to avoid bringing in a jQuery plugin, I chose to write
functions that would parse 12hr time as well as return the day of the
week in string format. I don't think it is a dirty hack and we currently
aren't bringing in any js libraries that enable this functionality, so I
think for now this is acceptable.