DE4551 Fix time issue
1. Contextualize any major architectural or UI changes in this PR.
This fixes a bug in which, when the minutes of the time was between 0
and 10, it would render a single digit to the web page, resulting in a
time like 10:7 am, instead of 10:07 am. This PR fixes that issue by
checking to see if the minutes of the start time with offset is less
than 10. If it is, it adds a '0' before that digit.

2. Do these changes belong here, or somewhere else?
Here

3. Does the code follow our coding standards and practices?
Yes

4. Has the code been linted?
Yes

5. Have you written tests for the code you wrote / changed?
N/A

6. Have you run your tests locally?
Yes

7. Is there anything that feels like a dirty hack? If so, have you worked with
someone to find another, cleaner way to accomplish the same thing?
Nope / N/A