US10205 Get location distances from origin
1. Contextualize any major architectural or UI changes in this PR.
As part of US10205, a new API endpoint of /locations/proximites is
exposed and returns distance data of an origin and different crossroads
locations.

This commit provides javascript code that makes a call to that new API
endpoint and formats the response into an array of javascript objects
that contain location names and their distance from origin. It then
stores this formatted response into the distancesFromOrigin property on
the LocationFinder instance.

2. Do these changes belong here, or somewhere else?
Here

3. Does the code follow our coding standards and practices?
Yes

4. Has the code been linted?
Yes

5. Have you written tests for the code you wrote / changed?
Yes

6. Have you run your tests locally?
Yes

7. Is there anything that feels like a dirty hack? If so, have you worked with
someone to find another, cleaner way to accomplish the same thing?
No / N/A