US10205 Unit Test DistanceSorter
1. Contextualize any major architectural or UI changes in this PR.
This commit finishes unit tests for DistanceSorter.

2. Do these changes belong here, or somewhere else?
Here

3. Does the code follow our coding standards and practices?
Yes

4. Has the code been linted?
Yes

5. Have you written tests for the code you wrote / changed?
Yes

6. Have you run your tests locally?
Yes

7. Is there anything that feels like a dirty hack? If so, have you worked with
someone to find another, cleaner way to accomplish the same thing?
Sort of. We had to create a new instance of CardCarousels in order to
avoid some _really_ difficult mocking. Consulted with Rick Barker to do
this. We both feel it is the best way to move forward with these unit
tests.