PrerenderIO for Maestro
Added nginx to the maestro container as a daemon. Traffic is forwarded to nginx first so it can handle prerender and then proxied to maestro (cowboy). I also cleaned up some unused stuff like the MAESTRO_SSL key as we are terminating ssl at the ingress controller not at maestro.